eliminate compiler warning introduced by recent changes
git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.2@43895 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ cli.c
@@ -784,7 +784,7 @@ static int handle_showchan(int fd, int argc, char *argv[])
 		c->fds[0], c->fin & 0x7fffffff, (c->fin & 0x80000000) ? " (DEBUGGED)" : "",
 		c->fout & 0x7fffffff, (c->fout & 0x80000000) ? " (DEBUGGED)" : "", (long)c->whentohangup,
 		cdrtime, c->_bridge ? c->_bridge->name : "<none>", ast_bridged_channel(c) ? ast_bridged_channel(c)->name : "<none>", 
-		c->context, c->exten, c->priority, c->callgroup, c->pickupgroup, ( c->appl ? c->appl : "(N/A)" ),
+		c->context, c->exten, c->priority, (int) c->callgroup, (int) c->pickupgroup, ( c->appl ? c->appl : "(N/A)" ),
 		( c-> data ? (!ast_strlen_zero(c->data) ? c->data : "(Empty)") : "(None)"),
 		(ast_test_flag(c, AST_FLAG_BLOCKING) ? c->blockproc : "(Not Blocking)"));
